py_version/requirements.txt
PyQt5~=5.15.7
PyAutoGUI~=0.9.53
pynput~=1.7.6
PyQtWebEngine~=5.15.6
SQLAlchemy~=1.4.44
opencv-python~=4.6.0.66
numpy~=1.23.5
matplotlib~=3.6.2
mss~=7.0.1
Pillow~=9.3.0